f64 to f32

+++ src/activators/types.rs
@@ -1,4 +1,4 @@
 use crate::maths::types::MatrixD;
 
-pub type Activator = fn(MatrixD<f64>) -> MatrixD<f64>;
-pub type ActivatorDeriv = fn(MatrixD<f64>) -> MatrixD<f64>;
+pub type Activator = fn(MatrixD<f32>) -> MatrixD<f32>;
+pub type ActivatorDeriv = fn(MatrixD<f32>) -> MatrixD<f32>;
